+++ SurfaceTopography/IO/BCR.py
@@ -28,20 +28,25 @@
 # https://sourceforge.net/p/gwyddion/code/HEAD/tree/trunk/gwyddion/modules/file/bcrfile.c
 #
 
+import logging
+
 import numpy as np
 
 from .common import OpenFromAny
-from ..Exceptions import CorruptFile, FileFormatMismatch, MetadataAlreadyFixedByFile
+from ..Exceptions import CorruptFile, FileFormatMismatch, MetadataAlreadyFixedByFile, UnsupportedFormatFeature
 from ..UniformLineScanAndTopography import Topography
-from ..Support.UnitConversion import get_unit_conversion_factor
+from ..Support.UnitConversion import get_unit_conversion_factor, is_length_unit
 
 from .Reader import ReaderBase, ChannelInfo
 
+_log = logging.getLogger(__file__)
 
-###
 
 class BCRReader(ReaderBase):
     _format = 'bcr'
+    _mime_types = ['application/x-bcr-spm', 'application/x-bcrf-spm']
+    _file_extensions = ['bcr', 'bcrf']
+
     _name = 'BCR-STM file format'
     _description = '''
 BCR-STM and BCRF file formats
@@ -125,11 +130,14 @@ BCR-STM and BCRF file formats
             self._metadata = {}
             eof = False
             while not eof:
-                if not line.startswith('%'):
-                    key, value = line.split('=', 1)
-                    key = key.strip()
-                    value = value.strip()
-                    self._metadata[key] = value
+                if not line.startswith('%') and not line.startswith('#'):
+                    try:
+                        key, value = line.split('=', 1)
+                        key = key.strip()
+                        value = value.strip()
+                        self._metadata[key] = value
+                    except ValueError:
+                        _log.warning(f"Skipping line '{line}' because it does not appear to be a key/value pair.")
                 try:
                     line, buffer_str = buffer_str.split('\n', 1)
                     line = line.strip()
@@ -144,6 +152,10 @@ BCR-STM and BCRF file formats
         xunit = self._metadata['xunit']
         assert xunit == self._metadata['yunit']
         zunit = self._metadata['zunit']
+
+        if not is_length_unit(zunit):
+            raise UnsupportedFormatFeature(f"This BCR/BCRF file reports data in units of '{zunit}'. This is not "
+                                           f"a height unit as expected for topography data.")
 
         self._channels = [
             ChannelInfo(self,